+++ src/dsp/PhaseOscillator.hpp
@@ -6,7 +6,7 @@
 namespace myrisa {
 
 struct PhaseOscillator {
-  double phase = 0.0f;
+  float phase = 0.0f;
   float freq = 1.0f;
 
   PhaseOscillator() {}
